@@ -15,15 +15,15 @@
     "systemjs-demo": "tsc -p src/tsconfig-es5-commonjs.json",
     "parcel-demo": "ngc -p src/tsconfig-es6.json && parcel build built-es6/src/apps/bundler-comparison-app/main.js --out-dir dist/parcel-bundle",
     "systemjs-builder-demo": "npm run systemjs-demo && node systemjs-builder/systemjs-builder.js",
-    "webpack-no-lazy-loading": "ngc -p src/tsconfig-es5.json && webpack --config webpack/webpack-no-lazy-loading.js --progress",
-    "webpack-lazy-loading": "webpack --config webpack/webpack-lazy-loading.js --progress",
-    "webpack-no-angular-optimizations": "ngc -p src/tsconfig-es5.json && webpack --config webpack/webpack-no-angular-optimizations.js --progress",
+    "webpack-no-lazy-loading": "ngc -p src/tsconfig-es5.json && webpack --config webpack/webpack-no-lazy-loading.js --progress --mode production",
+    "webpack-lazy-loading": "webpack --config webpack/webpack-lazy-loading.js --progress --mode production",
+    "webpack-no-angular-optimizations": "ngc -p src/tsconfig-es5.json && webpack --config webpack/webpack-no-angular-optimizations.js --progress --mode production",
     "webpack-babili": "ngc -p src/tsconfig-es6.json && webpack --config webpack/webpack-babili.js --progress",
     "build-all": "npm run aot-i18n-en && npm run aot-i18n-no && npm run rollup-lazy && npm run browserify-demo && npm run parcel-demo && npm run webpack-babili && npm run webpack-lazy-loading && npm run webpack-no-lazy-loading && npm run systemjs-builder-demo && npm run build-demo-app && npm run closure-compiler-demo && npm run rollup-demo && npm run rollup-es6"
   },
   "devDependencies": {
-    "@angular-devkit/build-optimizer": "0.0.36",
-    "@ngtools/webpack": "1.9.3",
+    "@angular-devkit/build-optimizer": "0.3.1",
+    "@ngtools/webpack": "6.0.3",
     "babili-webpack-plugin": "^0.1.1",
     "browserify": "14.5.0",
     "compression": "1.6.2",
@@ -32,40 +32,41 @@
     "path": "0.12.7",
     "raw-loader": "0.5.1",
     "systemjs-builder": "0.16.4",
-    "typescript": "2.4.2",
+    "typescript": "2.7.2",
     "uglify-js": "3.2.2",
-    "webpack": "3.10.0"
+    "webpack": "4.8.3",
+    "webpack-cli": "2.1.3"
   },
   "dependencies": {
+    "@angular/animations": "6.0.2",
+    "@angular/common": "6.0.2",
+    "@angular/compiler": "6.0.2",
+    "@angular/compiler-cli": "6.0.2",
+    "@angular/core": "6.0.2",
+    "@angular/forms": "6.0.2",
+    "@angular/http": "6.0.2",
+    "@angular/platform-browser": "6.0.2",
+    "@angular/platform-browser-dynamic": "6.0.2",
+    "@angular/platform-server": "6.0.2",
+    "@angular/router": "6.0.2",
+    "@angular/upgrade": "6.0.2",
     "angular": "~1.4.7",
-    "@angular/animations": "6.0.0-beta.1",
-    "@angular/common": "6.0.0-beta.1",
-    "@angular/compiler": "6.0.0-beta.1",
-    "@angular/compiler-cli": "5.2.1",
-    "@angular/core": "6.0.0-beta.1",
-    "@angular/forms": "6.0.0-beta.1",
-    "@angular/http": "6.0.0-beta.1",
-    "@angular/platform-browser": "6.0.0-beta.1",
-    "@angular/platform-browser-dynamic": "6.0.0-beta.1",
-    "@angular/platform-server": "6.0.0-beta.1",
-    "google-closure-compiler": "20171023.0.1",
-    "@angular/router": "6.0.0-beta.1",
-    "@angular/upgrade": "6.0.0-beta.1",
     "bootstrap": "~3.3.4",
     "core-js": "^2.4.1",
     "events": "^1.0.2",
     "fbjs": "^0.3.1",
     "flux": "^2.1.1",
-    "rollup": "0.55.0",
-    "rollup-plugin-commonjs": "8.2.6",
-    "rollup-plugin-uglify": "2.0.1",
-    "rollup-plugin-node-resolve": "3.0.0",
-    "rxjs": "5.5.6",
-    "systemjs": "0.20.19",
-    "zone.js": "0.8.6",
+    "google-closure-compiler": "20171023.0.1",
     "jquery": "3.3.1",
     "jquery-ui-dist": "1.12.1",
     "react": "~0.13.3",
-    "underscore": "~1.8.3"
+    "rollup": "0.55.0",
+    "rollup-plugin-commonjs": "8.2.6",
+    "rollup-plugin-node-resolve": "3.0.0",
+    "rollup-plugin-uglify": "2.0.1",
+    "rxjs": "6.1.0",
+    "systemjs": "0.20.19",
+    "underscore": "~1.8.3",
+    "zone.js": "0.8.26"
   }
 }
