@@ -1,11 +1,14 @@
-import {Subject } from 'rxjs/Subject';
-import {Coordinates} from './coordinates';
-import 'rxjs/add/operator/bufferCount';
+import { Subject, pipe } from "rxjs";
+import { Coordinates } from "./coordinates";
+import { bufferCount, map } from "rxjs/operators";
 
-export class EdgeService extends Subject<Coordinates>{
-
-    getCoordinates(){
-        return this.asObservable().bufferCount(2).map(buffer => {return {first:buffer[0], second:buffer[1]}});
-    }
-        
+export class EdgeService extends Subject<Coordinates> {
+  getCoordinates() {
+    return this.asObservable().pipe(
+      bufferCount(2),
+      map(buffer => {
+        return { first: buffer[0], second: buffer[1] };
+      })
+    );
+  }
 }
