@@ -1,5 +1,5 @@
 import {Component, OnInit} from '@angular/core';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 
 import {Store} from './store';
 import {LogAction} from './log-action';
@@ -7,8 +7,8 @@ import {LogEntry} from './log-entry';
 import {LogTailService} from './log-tail-service';
 
 @Component({
-    providers:[Store,LogTailService],
-    template:`<div>
+  providers: [Store, LogTailService],
+  template: `<div>
                 <h1>Error log state managed using Redux</h1>
                 <button style="margin-bottom: 10px;" (click)="generateLogEntry()">Add new log entry</button>
                 <input placeholder="message" type="text" [(ngModel)]="msg" />
@@ -23,29 +23,31 @@ import {LogTailService} from './log-tail-service';
                      </tr>
                 </table>
                 <h4><a href="http://www.syntaxsuccess.com/viewarticle/redux-in-angular-2.0">Read more here</a></h4>
-              </div>`
+              </div>`,
 })
+export class LogDemo implements OnInit {
+  store: Store;
+  msg: string;
+  severity: number;
+  logTailService: LogTailService;
 
-export class LogDemo implements OnInit{
+  constructor(store: Store, logTailService: LogTailService) {
+    this.store = store;
+    this.logTailService = logTailService;
+  }
 
-    store:Store;
-    msg:string;
-    severity:number;
-    logTailService:LogTailService
+  generateLogEntry() {
+    let entry = new LogEntry(this.msg, this.severity);
+    this.store.dispatchAction(new LogAction('ADD_ENTRY', entry));
+    this.msg = '';
+    this.severity = undefined;
+  }
 
-    constructor(store:Store, logTailService:LogTailService){
-        this.store = store;
-        this.logTailService = logTailService;
-    }
-
-    generateLogEntry(){
-        let entry = new LogEntry(this.msg, this.severity);
-        this.store.dispatchAction(new LogAction('ADD_ENTRY',entry));
-        this.msg = '';
-        this.severity = undefined;
-    }
-
-    ngOnInit(){
-        this.logTailService.getLogEntries().subscribe((res) => this.store.dispatchAction(new LogAction('LOAD_ENTRIES', res.entries)));
-    }
+  ngOnInit() {
+    this.logTailService
+      .getLogEntries()
+      .subscribe(res =>
+        this.store.dispatchAction(new LogAction('LOAD_ENTRIES', res.entries)),
+      );
+  }
 }
