@@ -3,8 +3,6 @@ import {CommonModule} from '@angular/common';
 import {RouterModule} from '@angular/router';
 import {HttpModule} from '@angular/http';
 
-import 'rxjs/add/operator/map';
-
 import {CountryDemo} from '../../shared-components/lazy-loaded-tree-view/country-demo';
 import {LazyTreeView} from '../../shared-components/lazy-loaded-tree-view/tree-view';
 
@@ -13,10 +11,7 @@ import {LazyTreeView} from '../../shared-components/lazy-loaded-tree-view/tree-v
   imports: [
     CommonModule,
     HttpModule,
-    RouterModule.forChild(
-    [
-      {path: '', component: CountryDemo}
-    ])
-  ]
+    RouterModule.forChild([{path: '', component: CountryDemo}]),
+  ],
 })
 export class LazyTreeviewModule {}
