@@ -1,35 +1,15 @@
-const UglifyJsPlugin = require('webpack/lib/optimize/UglifyJsPlugin');
+const UglifyJsPlugin = require('uglifyjs-webpack-plugin');
+const uglifyOptions = require('./uglify-options');
 const webpack = require('webpack');
 
 module.exports = {
-  entry: './built-es5/app/main.js',
+  entry: './built-es5/src/apps/bundler-comparison-app/main.js',
   output: {
     filename: 'dist/bundle-webpack-3-no-angular-optimizations.js'
   },
 
   plugins: [
     new webpack.optimize.ModuleConcatenationPlugin(),
-    new UglifyJsPlugin({
-      beautify: false, 
-      output: {
-        comments: false
-      }, 
-      mangle: {
-        screw_ie8: true
-      }, 
-      compress: {
-        screw_ie8: true,
-        warnings: false,
-        conditionals: true,
-        unused: true,
-        comparisons: true,
-        sequences: true,
-        dead_code: true,
-        evaluate: true,
-        if_return: true,
-        join_vars: true,
-        negate_iife: false
-      },
-    })
+    new UglifyJsPlugin({uglifyOptions:uglifyOptions})
   ]
 }
